third_party/WebKit/LayoutTests/crypto/subtle/algorithm-identifier-as-string-expected.txt

Tests that a string can be passed in place of an AlgorithmIdentifier

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS key.type is "secret"
PASS key.algorithm.name is "AES-CBC"
PASS key.algorithm.length is 128
PASS: sha-1 of empty string should be [da39a3ee5e6b4b0d3255bfef95601890afd80709] and was
PASS key.type is "secret"
PASS key.algorithm.name is "HMAC"
PASS key.algorithm.hash.name is "SHA-256"
PASS key.algorithm.length is 120
PASS successfullyParsed is true

TEST COMPLETE
